pyproject.toml
[project]
name = "mopbot"
version = "0.1.0"
description = ""
authors = [
    {name = "Morpheus636",email = "morpheus636@morpheus636.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "discord-py (>=2.6.4,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
